# frontend/src/app/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Suivie Physique API",
    "description": "API for Suivie Physique Application",
    "termsOfService": "Terms of service",
    "contact": { "name": "Larbi Dotcipher", "email": "dotcipher@ncrm.com" },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:3008/api/v1", "description": "Local server" },
    {
      "url": "https://suivie-physique.herokuapp.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/users/change-status": {
      "patch": {
        "tags": ["users-controller"],
        "operationId": "changeStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangeStatusRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/users/change-role": {
      "patch": {
        "tags": ["users-controller"],
        "operationId": "changeRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangeRoleRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/users/change-password": {
      "patch": {
        "tags": ["users-controller"],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/users/change-member": {
      "patch": {
        "tags": ["users-controller"],
        "operationId": "changeMember",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangeMemberRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/users/change-email": {
      "patch": {
        "tags": ["users-controller"],
        "operationId": "changeEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangeEmailRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeStatusRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "newStatus": { "type": "boolean" }
        }
      },
      "ChangeRoleRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "newRole": { "type": "string" }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "oldPassword": { "type": "string" },
          "newPassword": { "type": "string" },
          "confirmPassword": { "type": "string" }
        }
      },
      "ChangeMemberRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "newRole": { "type": "string" },
          "newStatus": { "type": "boolean" }
        }
      },
      "ChangeEmailRequest": {
        "type": "object",
        "properties": { "newEmail": { "type": "string" } }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}